+++ winim/extra.nim
@@ -7,8 +7,6 @@
 #
 #====================================================================
 
-{.deadCodeElim: on.}
-
 import inc/[psapi, tlhelp32, wintrust, imagehlp, odbc, dde, lzexpand, nb30,
   winperf, mmsystem, winefs, winioctl, winscard, winspool, mcx, powrprof, vfw,
   usp10, subauth, snmp, aclapi, dhcpcsdk, wincodec, gdiplus]