+++ winim/mean.nim
@@ -5,7 +5,5 @@
 #
 #====================================================================
 
-{.deadCodeElim: on.}
-
 import utils, winstr, core, shell, ole
 export utils, winstr, core, shell, ole