+++ winim/net.nim
@@ -7,7 +7,5 @@
 #
 #====================================================================
 
-{.deadCodeElim: on.}
-
 import inc/[security, windns, winldap, lm, ras, rpc, iphlpapi, http, wincred]
 export security, windns, winldap, lm, ras, rpc, iphlpapi, http, wincred