+++ winim/winstr.nim
@@ -155,9 +155,6 @@
 ##    proc newWideCString(s: wstring): WideCString
 ##      # Converts wstring to WideCString.
 
-{.push hint[Name]: off.}
-{.deadCodeElim: on.}
-
 import macros, strutils, inc/[winimbase, windef]
 export strutils.toHex, winimbase